+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1507,7 +1507,7 @@ class Evaluations3(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-06-26T10:43:48.119Z")
+    ] = parse_datetime("2026-06-26T14:08:00.232Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1652,7 +1652,7 @@ class Evaluations2(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-06-26T10:43:48.118Z")
+    ] = parse_datetime("2026-06-26T14:08:00.231Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1797,7 +1797,7 @@ class Evaluations1(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-06-26T10:43:48.117Z")
+    ] = parse_datetime("2026-06-26T14:08:00.230Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1902,7 +1902,7 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-26T10:43:25.088Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-26T14:07:38.161Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")