+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KW1REMNGYNBDPJ1DEAYBCK6D"
+        "01KW244HK2WKZSBJJ3N0WK450T"
     )
     r"""The unique identifier of the data source"""