+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KW1REMNET9RFZ7CNMZ9BMJTC"
+        "01KW244HK00ATWGNPWPNS0AWNY"
     )
     r"""The unique identifier of the data source"""