+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KW1REMNM9K8BZ21RVNRX0T46"
+        "01KW244HK95TPV3CZGZX4JV6KD"
     )
     r"""The unique identifier of the data source"""